Add no-op implementation of logger. Fix typo in package naming.
