+++ dbinspector-no-op/proguard-rules.txt
@@ -1,3 +1,16 @@
 -keep public class com.infinum.dbinspector.DbInspector {
   public protected *;
 }
+-keep public class com.infinum.dbinspector.data.models.memory.logger.Level {
+  public protected *;
+}
+-keep public class com.infinum.dbinspector.data.sources.memory.logger.Logger {
+  public protected *;
+}
+-keep public class com.infinum.dbinspector.data.sources.memory.logger.EmptyLogger {
+  public protected *;
+}
+-keep public class com.infinum.dbinspector.data.sources.memory.logger.AndroidLogger {
+  public protected *;
+}
+
