+++ dbinspector-no-op/src/main/kotlin/com/infinum/dbinspector/DbInspector.kt
@@ -1,7 +1,11 @@
 package com.infinum.dbinspector
 
+import com.infinum.dbinspector.data.sources.memory.logger.EmptyLogger
+import com.infinum.dbinspector.data.sources.memory.logger.Logger
+
+@Suppress("UnusedPrivateMember")
 public object DbInspector {
 
     @JvmStatic
-    public fun show(): Unit = Unit
+    public fun show(logger: Logger = EmptyLogger()): Unit = Unit
 }
