+++ dbinspector-no-op/src/main/kotlin/com/infinum/dbinspector/data/models/memory/logger/Level.kt
@@ -1,0 +1,8 @@
+package com.infinum.dbinspector.data.models.memory.logger
+
+public enum class Level {
+    DEBUG,
+    INFO,
+    ERROR,
+    NONE
+}
