+++ dbinspector/src/main/kotlin/com/infinum/dbinspector/data/Data.kt
@@ -7,17 +7,17 @@ import com.infinum.dbinspector.data.Data.Constants.Name.PROTO_FILENAME_HISTORY
 import com.infinum.dbinspector.data.Data.Constants.Name.PROTO_FILENAME_SETTINGS
 import com.infinum.dbinspector.data.models.local.proto.output.HistoryEntity
 import com.infinum.dbinspector.data.models.local.proto.output.SettingsEntity
-import com.infinum.dbinspector.data.source.local.cursor.PragmaSource
-import com.infinum.dbinspector.data.source.local.cursor.RawQuerySource
-import com.infinum.dbinspector.data.source.local.cursor.SchemaSource
-import com.infinum.dbinspector.data.source.local.proto.history.HistoryDataStore
-import com.infinum.dbinspector.data.source.local.proto.history.HistorySerializer
-import com.infinum.dbinspector.data.source.local.proto.settings.SettingsDataStore
-import com.infinum.dbinspector.data.source.local.proto.settings.SettingsSerializer
-import com.infinum.dbinspector.data.source.memory.connection.AndroidConnectionSource
-import com.infinum.dbinspector.data.source.memory.pagination.CursorPaginator
-import com.infinum.dbinspector.data.source.memory.pagination.Paginator
-import com.infinum.dbinspector.data.source.raw.AndroidDatabasesSource
+import com.infinum.dbinspector.data.sources.local.cursor.PragmaSource
+import com.infinum.dbinspector.data.sources.local.cursor.RawQuerySource
+import com.infinum.dbinspector.data.sources.local.cursor.SchemaSource
+import com.infinum.dbinspector.data.sources.local.proto.history.HistoryDataStore
+import com.infinum.dbinspector.data.sources.local.proto.history.HistorySerializer
+import com.infinum.dbinspector.data.sources.local.proto.settings.SettingsDataStore
+import com.infinum.dbinspector.data.sources.local.proto.settings.SettingsSerializer
+import com.infinum.dbinspector.data.sources.memory.connection.AndroidConnectionSource
+import com.infinum.dbinspector.data.sources.memory.pagination.CursorPaginator
+import com.infinum.dbinspector.data.sources.memory.pagination.Paginator
+import com.infinum.dbinspector.data.sources.raw.AndroidDatabasesSource
 import com.infinum.dbinspector.extensions.dataStoreFile
 import org.koin.core.module.Module
 import org.koin.core.qualifier.StringQualifier
