--- dbinspector/src/main/kotlin/com/infinum/dbinspector/data/source/memory/logger/EmptyLogger.kt
@@ -1,8 +1,0 @@
-package com.infinum.dbinspector.data.source.memory.logger
-
-import com.infinum.dbinspector.data.models.memory.logger.Level
-
-internal class EmptyLogger : Logger(Level.NONE) {
-
-    override fun log(level: Level, message: String) = Unit
-}
