+++ dbinspector/src/main/kotlin/com/infinum/dbinspector/data/sources/local/proto/settings/SettingsDataStore.kt
@@ -1,0 +1,20 @@
+package com.infinum.dbinspector.data.sources.local.proto.settings
+
+import androidx.datastore.core.DataStore
+import com.infinum.dbinspector.data.Sources
+import com.infinum.dbinspector.data.models.local.proto.output.SettingsEntity
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.firstOrNull
+
+internal class SettingsDataStore(
+    private val store: DataStore<SettingsEntity>
+) : Sources.Local.Settings {
+
+    override suspend fun store(): DataStore<SettingsEntity> = store
+
+    override suspend fun current(): SettingsEntity =
+        store.data.firstOrNull() ?: SettingsEntity.getDefaultInstance()
+
+    override fun flow(): Flow<SettingsEntity> =
+        store.data
+}
