+++ dbinspector/src/main/kotlin/com/infinum/dbinspector/data/sources/memory/logger/EmptyLogger.kt
@@ -1,0 +1,8 @@
+package com.infinum.dbinspector.data.sources.memory.logger
+
+import com.infinum.dbinspector.data.models.memory.logger.Level
+
+internal class EmptyLogger : Logger(Level.NONE) {
+
+    override fun log(level: Level, message: String) = Unit
+}
