+++ dbinspector/src/main/kotlin/com/infinum/dbinspector/ui/Presentation.kt
@@ -2,7 +2,7 @@ package com.infinum.dbinspector.ui
 
 import android.annotation.SuppressLint
 import android.content.Context
-import com.infinum.dbinspector.data.source.memory.logger.Logger
+import com.infinum.dbinspector.data.sources.memory.logger.Logger
 import com.infinum.dbinspector.di.LibraryKoin
 import com.infinum.dbinspector.domain.Domain
 import com.infinum.dbinspector.ui.content.table.TableViewModel
