+++ sample/build.gradle
@@ -34,6 +34,7 @@ android {
         debug {
             debuggable true
             minifyEnabled false
+            applicationIdSuffix ".debug"
         }
         release {
             debuggable false
