+++ sample/src/main/kotlin/com/infinum/dbinspector/sample/MainActivity.kt
@@ -3,7 +3,7 @@ package com.infinum.dbinspector.sample
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.infinum.dbinspector.DbInspector
-import com.infinum.dbinspector.data.source.memory.logger.AndroidLogger
+import com.infinum.dbinspector.data.sources.memory.logger.AndroidLogger
 import com.infinum.dbinspector.sample.databinding.ActivityMainBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
@@ -13,6 +13,8 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        supportActionBar?.title = "${supportActionBar?.title?.toString().orEmpty()} (${BuildConfig.BUILD_TYPE})"
 
         ActivityMainBinding.inflate(layoutInflater)
             .also { setContentView(it.root) }
